修复
未知

+++ HHCommonUtils/HHCommonUtils/HHCommonUtils.m
@@ -794,10 +794,10 @@
     return standardStringValue;
 }
 
-+ (NSString *)floatString:(NSString*)arg fromDictionary:(NSDictionary*)dictionary tailLength:(NSUInteger)tailLength{
++ (NSString *)floatString:(NSString*)arg fromDictionary:(NSDictionary*)dictionary tailLength:(NSInteger)tailLength{
     NSString *standardStringValue = [self stringValue:arg fromDictionary:dictionary];
     NSString *headString = @"%";
-    NSString *willAppendString = [NSString stringWithFormat:@".%ldf",(long)tailLength];
+    NSString *willAppendString = [NSString stringWithFormat:@".%ldf",tailLength];
     NSString *formatString = [headString stringByAppendingString:willAppendString];
     NSScanner* scan = [NSScanner scannerWithString:standardStringValue];
     float val;
